// frontend/src/api/model/normativeLevel3NodeDto.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { NormativeIndicatorDto } from './normativeIndicatorDto.ts';

export interface NormativeLevel3NodeDto {
  id?: string;
  name?: string;
  label?: string;
  order?: number;
  description?: string;
  indicators?: NormativeIndicatorDto[];
}